chore: Upgrade Hyper API to 0.0.22502 (#76)
In particular, this upgrade includes support for two new parameters:
* `QueryInfoParam.schema_output_format` allows to fetch the schema using
   a query-id, even if the query produced zero rows, and hence also zero
   chunks. This will be required for the Spark driver to retrieve the
   schema of a query result given a query-result-id.
* `QueryParam.query_row_limit` allows to fetch the result of a SQL query
   without actually running the query. This will be required for submitting
   a Hyper-SQL query from the Spark-SQL connector.

Furthermore, with this update we are also including all comments in the
`.proto` files. This will make it easier to improve the JDBC driver in
the future without always having to keep the internal version of this
file open in a separate tab.