debt: Refactor Property Handling (#77)
In preparation for cleaning up the various timeout handlings in the JDBC
driver this change first cleans up the property handling.
All properties will now be interpreted when the connection is created
instead of the previous approach where it was more distributed at the
places where the respective properties are used.
For queryTimeout this for example had the benefit that it's much easier
to understand how it is computed and the internal code doesn't need to
check for invalid user inputs again and again.

The property handling is now centralized into `...Properties` classes
(like `ConnectionProperties` and `StatementProperties`) that coarsely
match the area for which they are relevant.
This structuring helped to for example identify that
`resultset.bytelimit` was used at the wrong levels (see notes further
down).
For now the new approach only covers properties relevant for the
Connection & Statement objects.
_Future_ refactorings should also make use of it for the Authorization
related properties.
With the transition of properties to this new framework we also removed
the "silent" ignoring of malformed property values.
Just ignoring property values for valid keys can lead to surprising
behavior like for example timeouts getting ignored, thus explicitly
failing is beneficial.

Removed Properties:
- The `resultset.bytelimit` property was removed. It didn't work for
usage in DBeaver or other generic JDBC clients anyway as it would only
have effect when the max rows are limited. As currently it is sufficient
that these settings can be configured through our API extensions instead
of fixing and adding more properties for maxRows this change removes the
property.

Related Changes:
- Double enforcement of query timeout on stub and in the executor: The
query timeout was used both for individual generic RPC requests at the
connection level as well as governing the whole end to end query
execution flow encompassing multiple RPC requests. To clean this up the
individual RPC call timeout is now detached from the queryTimeout and
instead associated with the networkTimeout (with the standardized get /
set support).
- Don't keep `byteLimit` as a member in `HyperGprcClientExecutor`. The
`byteLimit` is only relevant when there is a row limit. In the existing
code the byteLimit was set as a member during construction for all use
cases while the row limit was only passed as argument in specific
`executeQuery` overloads. This split was highly confusing. The
`byteLimit` is now explicitly passed in on the methods that need it
alongside the row limit.
- This change fixes the behavior of the `client info` related connection
methods. Our driver currently doesn't support client info properties and
thus we just return an empty properties set / null when returning those
and ignore set values. This behavior mirrors the documented behavior of
the MSSQL driver.
- Removed support for legacy `serverSetting.` prefix

The actual proper timeout handling alignment and cleanup will be done
separately in a follow up.