feat: Improved query timeout handling (#80)
This change adjusts the query timeout handling from being locally + gRPC
deadline enforced to now primarily being enforced on Hyper server side.
- This allows to return the server side error to the caller. Previously
it would either raise an internal error by the JDBC driver or obscure
gRPC timeout errors depending on which layer detected the problem first.
- It now also guarantees that async queries don't exceed their timeout
even if the client crashes / or the user flow forgets to explicitly
invoke `cancel`.

To ensure robustness of the query timeout handling in face of network
problems or even server side bugs this change doesn't only rely on the
server side timeouts.
As a safety net there is also local enforcement (which happens with a
configurable delay to allow the server to terminate first)

As a related change the driver now also internally tracks timeouts more
accurately (both for executeQuery as well as for the custom `wait*`
methods).
- The old approach was in some places relying on non monotonic clocks
(`Instant`) and thus would have resulted in failures around system clock
updates / summer time changes
- The old approach sometimes waited for a multiple of the configured
query timeout

To remove this class of issues usages a new Deadline class is introduced
that tracks the remaining time for a timeout.