fix: Restrict Avatica to version before 1.27.0 to make the driver buildable again (#117)
The 1.27.0 Avatica updated broke our
`jdbc-core/src/main/java/com/salesforce/datacloud/jdbc/core/accessor/QueryJDBCAccessor.java`
class as adjusted the `Cursor` interface to add support for MySQL
unsigned numbers https://issues.apache.org/jira/browse/CALCITE-5094.

This is not something that is of use to us and thus with the upcoming
remove of Avatica in mind this change just pins the version to a release
that doesn't have this change yet.