[chore] Standardize on `SQLException` in our code base (#124)
Before this change `DataCloudJDBCException` and `SQLException` were used
inconsistently in our public interfaces. With this change the public
interfaces now consistently use `SQLException`.

Additionally the use of `DataCloudJDBCException` was unnecessary in many
places as it was just delegating to `SQLException`. All those places got
adjusted and with that also the `throws` declarations / test
expectations of users of those functions.