feat: Allow to toggle the inclusion of query text details in exception messages (#126)
For some use cases like logging in an app server it can be undesirable
to log query / snippets as they might contain data that should not be
logged. This change adds new property `errorsIncludeCustomerDetails`
that allows removing such information from the generic exception
message. The default is kept like before to still show it. Independent
of the mode the full information is preserved in the exception and can
be access using `getFullCustomerMessage()`.

As the change needs to no support having the query included / excluded
from the message the format was adjusted to ensure such an
inclusion/exclusion doesn't impact the readability of the message (same
also for query id, which might be available in some context and missing
in others). The new general error formatting is
```
Failed to execute query: <server side message / local exception message>
SQLSTATE: <server provided / "HY==" for generic error>
QUERY-ID: ... <-- Line only if available
DETAIL: ... <-- Line only if available
HINT:... <-- Line only if available
QUERY: ... <-- Line only if available
SYSTEM-DETAIL:... <-- Line only if available
```

With this change the signature for exception creation was adjusted to
explicitly required always passing `query-id` and `query`. There were
many places where the query id could have been passed but wasn't and
same for the query. (Not all places were fixed in this PR because that
would require larger refactorings).

Drive-By:
- Remove unused signatures and their test coverage

Future work:
- The new interface uncovered several places on the execute query flow
where the query is not available for formatting the exception. Those
places should be adjusted so that they get access to the query string
- The error info contains additional information that is not yet part of
the exception. We should revisit our exception to add this information
as well.