fix: Remove required dataspace parameter from Connection (#133)
Previously the DataCloudConnection was special casing `dataspace` so
that it was required to be passed as a non-null construction parameter.

Upon further inspection, it turned out that this parameter is not needed
at all:
- For the OAuth based flows it's not required as dataspace is encoded in
the JWT
- For mTLS based flows, dataspace headers will be handled via
interceptors on the custom channel. Furthemore, the current
implementation is not sufficient anyway, as additional headers would be
needed, and those can currently only injected via interceptors, anyway.
 
Thus there is no scenario that the `dataspace` parameter enables on its
own. Hence, the passing of dataspace is removed.